Changes by dyllon
